requirements.txt
googlemaps==4.2.0
numpy==1.18.2
PyYAML==5.3.1
requests=2.23.0
beautifulsoup4==4.9.0
google==2.0.3
